requirements.txt
djangorestframework
aliyun-python-sdk-core-v3==2.8.9
aliyun-python-sdk-dysmsapi==1.0.0
bleach==2.1.3
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0.5
django-shortuuidfield==0.1.3
Flask==1.0.2
html5lib==1.0.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mysqlclient==1.3.12
Pillow==5.1.0
PyMySQL==0.8.1
python-memcached==1.59
pytz==2018.4
redis==2.10.6
requests==2.19.1
shortuuid==0.5.0
six==1.11.0
urllib3==1.23
webencodings==0.5.1
Werkzeug==0.14.1
